Updated the map, and added checking functionality
Updated the map, and added the functionality to allow the search to check for obstacles. The code is added to each move. before returning the move result it checks if the node is in an obstacle. If it is, it returns a None. The check obstacle function needs to just return a true or false if the node is in an obstacle or not. Good luck on the test. Text me if you have questions